adding s3 reading options
